🔧 #41: Encapsulate syncing UI
This commit refactors syncing the UI via AppleScript into an
encapsulated function. This function gets invoked when the Overlay
window is shown, but can also now by invoked at any time.

This helps to prepare the app for controlling the Music player, since we
will need a way to keep the UI in sync when the player changes state.