🐞 #75: Remove menu item shortcut
This commit fixes a bug where the menu item to toggle the overlay window
had a hard-coded shortcut, which could move out of sync with the user's
global hotkey setting.

This commit removes the hard-coded shortcut. SwiftUI does not currently
have a way to dynamically keep a menu item in sync with some variable.
And I'm not willing to overengineer an AppKit solution.

Closes #75.