🐞 #87: Omnibox character limit
This commit obviates a bug where autocomplete suggestions for long input
strings would be misaligned.

This happened because suggestions use a simple HStack for alignment,
while the TextField can scroll. Getting the two to line up perfectly in
longer strings has no trivially simple solution. It is also likely that
99% of user input will not be long enough to encounter the bug, so a
complex solution isn't warranted.

This commit gets around the limitation by introducing a character limit
for the omnibox. This prevents longer strings (that would scroll) from
existing. This can be easily replaced with a more robust solution if the
need arises.

Closes #87.